backend/requirements.txt
Flask
Flask-CORS
pandas
playwright==1.31.1
googlesearch-python==1.1.0